pyproject.toml
[tool.poetry]
name = "epc-data-tools"
version = "0.1.0"
description = "Tools for downloading and cleaning EPC data."
authors = ["Conor Dempsey <conor.dempsey@london.gov.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "epctools"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
python-dotenv = "^0.21.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
download-epc-data = "epctools.download_datasets:cli"
